==> src/preprocess/1_preprocess.ipynb <==
import pandas as pd
import os
import ast

base_dir = os.getcwd()  # Get the script's directory
output_dir = os.path.join(base_dir, "..", "output")

file_path1 = os.path.join(output_dir, "telegram_messages.xlsx")
file_path2 = os.path.join(output_dir, "telegram_media.xlsx")

# Load the DataFrames
df_text = pd.read_excel(file_path1)
df_media = pd.read_excel(file_path2)

df_text.head()

df_media.head()

# Drop redundant columns (Channel, Date, Media File) from review_df
df_media.drop(columns=['Channel', 'Date', 'Media File'], inplace=True)

# Merge the DataFrames
df_merged = pd.merge(df_text, df_media, on='Message_ID', how='left')

df_merged['Date'] = pd.to_datetime(df_merged['Date'])
df_merged.rename(columns={'Comments_x': 'Text Comments',
                          'Comments_y': 'Media Comments',
                          'Reactions_x': 'Text Reactions',
                          'Reactions_y': 'Media Reactions'}, inplace=True)

df_merged = df_merged[['Message_ID', 'Date', 'Message', 'Text_Comments', 'Text_Reactions', 'Media File', 'Media_Comments', 'Media_Reactions']]
df_merged.sample(10)

# Replace 'Text_Comments' and 'Text_Reactions' with blank where 'Message' is NaN
df_merged.loc[df_merged['Message'].isna(), ['Text_Comments', 'Text_Reactions']] = ""

# Replace 'Media_Comments' and 'Media_Reactions' with blank where 'Media File' is NaN
df_merged.loc[df_merged['Media File'].isna(), ['Media_Comments', 'Media_Reactions']] = ""

# Replace 'Text_Comments' with blank where 'Text_Comments' is NaN or '[]'
df_merged.loc[df_merged['Text_Comments'].isna() | (df_merged['Text_Comments'] == "[]"), 'Text_Comments'] = ""

# Replace 'Media_Comments' with blank where 'Media_Comments' is NaN or '[]'
df_merged.loc[df_merged['Media_Comments'].isna() | (df_merged['Media_Comments'] == "[]"), 'Media_Comments'] = ""


# Function to convert JSON-like string to dictionary safely
def parse_reactions(reaction_str):
    try:
        return ast.literal_eval(reaction_str) if isinstance(reaction_str, str) else {}
    except (SyntaxError, ValueError):
        return {}

# Apply conversion to both reaction columns
df_merged['Text_Reactions'] = df_merged['Text_Reactions'].apply(parse_reactions)
df_merged['Media_Reactions'] = df_merged['Media_Reactions'].apply(parse_reactions)

# Positive Emojis
positive_emojis = [
    "👍", "🥰", "🕊", "🎉", "🍾", "🏆", "💯", "🤩", "😍", "💘", "💋", "🤗",
    "🤝", "😎", "😇", "🦄", "🐳", "😴", "👨‍💻", "🎄", "🎅"
]

# Negative Emojis
negative_emojis = [
    "🤬", "🔥", "💔", "🤮", "🖕", "👎", "🤡", "🤯", "😈", "😱", "😡", "💩",
    "🤪", "💅", "🤡"
]

# Aggregated columns for Text reactions
df_merged["Text_Positive_Reactions"] = df_merged.apply(
    lambda row: sum(row.get('Text_Reactions', {}).get(f"{emoji}", 0) for emoji in positive_emojis), axis=1
)
df_merged["Text_Negative_Reactions"] = df_merged.apply(
    lambda row: sum(row.get('Text_Reactions', {}).get(f"{emoji}", 0) for emoji in negative_emojis), axis=1
)

# Aggregated columns for Media reactions
df_merged["Media_Positive_Reactions"] = df_merged.apply(
    lambda row: sum(row.get('Media_Reactions', {}).get(f"{emoji}", 0) for emoji in positive_emojis), axis=1
)
df_merged["Media_Negative_Reactions"] = df_merged.apply(
    lambda row: sum(row.get('Media_Reactions', {}).get(f"{emoji}", 0) for emoji in negative_emojis), axis=1
)

df_merged.drop(columns=['Text_Reactions', 'Media_Reactions'], inplace=True)
df_merged = df_merged[['Message_ID', 'Date', 'Message', 'Text_Comments', 'Text_Positive_Reactions', 'Text_Negative_Reactions',
                       'Media File', 'Media_Comments', 'Media_Positive_Reactions', 'Media_Negative_Reactions']]

df_merged.sample(5)

# Exploding the 'Text_Comments' column
df_merged_exploded = df_merged.explode('Text_Comments')

# Exploding the 'Media_Comments' column
df_merged_exploded = df_merged_exploded.explode('Media_Comments')

file_path3 = os.path.join(output_dir, "preprocessed_data.xlsx")
df_merged_exploded.to_excel(file_path3, index=False)
