==> src/preprocess/3_transcribe.ipynb <==
import pandas as pd
import os
import re
import whisper

model = whisper.load_model("base")

# Global variable
count = 0

# Function to extract speech text from a video

def extract_text_from_video(video_path):
    global count
    try:
        result = model.transcribe(video_path)
        count += 1
        print((count/512)*100, "%")
        return result["text"]
    except Exception as e:
        print(f"Error processing {video_path}: {e}")
        return ""

base_dir = os.getcwd()  # Get the script's directory
output_dir = os.path.join(base_dir, "..", "..", "output")

file_path = os.path.join(output_dir, "data_with_OCR_text.xlsx")
df = pd.read_excel(file_path)

# Apply Whisper transcription only to MP4 files
df['Extracted_Text_from_Video'] = df['Media File'].apply(
    lambda x: extract_text_from_video(x) if pd.notna(x) and x.lower().endswith('.mp4') else ""
)

file_path1 = os.path.join(output_dir, "data_with_Whisper_text.xlsx")
df.to_excel(file_path1, index=False)

df.sample(5)
