==> src/preprocess/4_translate.ipynb <==
import pandas as pd
import os
from googletrans import Translator

def translate_to_english(text):
    try:
        # Translate to English
        translated = translator.translate(text, src='auto', dest='en').text
        return translated
    except Exception as e:
        return text  # If translation fails, return the original text

base_dir = os.getcwd()  # Get the script's directory
output_dir = os.path.join(base_dir, "..", "..", "output")

file_path = os.path.join(output_dir, "data_with_Whisper_text.xlsx")
df = pd.read_excel(file_path)

# Initialize the translator
translator = Translator()

df['Translated_Text'] = df['Extracted_Text_from_Video'].apply(translate_to_english)

df.drop(columns=['Extracted_Text_from_Video'], inplace=True)

file_path1 = os.path.join(output_dir, "final_preprocessed_data.xlsx")
df.to_excel(file_path1, index=False)
