==> src/preprocess/5. auto_label.ipynb <==
import pandas as pd
import os
from transformers import pipeline

base_dir = os.getcwd()  # Get the script's directory
output_dir = os.path.join(base_dir, "..", "..", "output")

file_path1 = os.path.join(output_dir, "data_with_OCR_text.xlsx")
file_path2 = os.path.join(output_dir, "final_preprocessed_data.xlsx")
df1 = pd.read_excel(file_path1)
df2 = pd.read_excel(file_path2)

df2 = df2.merge(df1[['Message_ID', 'Cleaned_Extracted_Text']], on='Message_ID', how='left')

df2["Combined_text"] = df2["Message"].fillna('') + " " + df2["Cleaned_Extracted_Text"].fillna('') + " " + df2["Translated_Text"].fillna('')

classifier = pipeline("zero-shot-classification", model="facebook/bart-large-mnli")
labels = ["news", "non-news"]

df2["Classification"] = df2["Combined_text"].apply(
    lambda x: classifier(x, labels)["labels"][0] if isinstance(x, str) and len(x.strip()) > 0 else "Unknown"
)

df2['Total_Comments'] = df2['Text_Comments'].fillna("") + df2['Message'].fillna("")

df2 = df2[df2['Classification'] != 'non-news']

df2.drop(columns=['Message', 'Text_Comments', 'Media File',
                  'Media_Comments', 'Translated_Text', 'Cleaned_Extracted_Text', 'Classification',
                  'Media_Positive_Reactions', 'Media_Negative_Reactions'], inplace=True)

cols = ['Message_ID', 'Date',  'Combined_text', 'Total_Comments', 'Text_Positive_Reactions', 'Text_Negative_Reactions']
df2 = df2[cols]

df2.head()

file_path1 = os.path.join(output_dir, "auto_labeled_data.xlsx")
df2.to_excel(file_path1, index=False)
